Simplified make, converted from jlc.py to jlc, and removed llvm-strip
It's now a single make target for each benchmark and the compiler
to be used and its flags are set in the Makefile.sub located in the
POLYBENCH_ROOT directory.

The python script jlc.py has been removed and jlc from jlm is now
used to compile the benchmarks.

The llvm-strip tool did not provide any benefit for performance
evaluations so it has been removed.